+++ semana12/labeddit/public/index.html
@@ -25,6 +25,7 @@
       html,
       body,
       #root {
+        background-color: #DAE0E6;
         margin: 0;
         padding: 0;
         width: 100%;
