+++ semana12/labeddit/src/hooks/useRequestData.js
@@ -22,4 +22,4 @@ const useRequestData = (initialData, url) => {
     return (data)
 }
 
-export default useRequestData
+export default useRequestData;
