+++ semana12/labeddit/src/pages/PostDetailsPage/PostDetailsPage.js
@@ -1,13 +1,20 @@
 import React from "react";
-import useProtectedPage from "../../hooks/useProtectedPage"
+import { useParams } from "react-router-dom"
+import useProtectedPage from "../../hooks/useProtectedPage";
+import Post from "../../components/Post/Post";
+
+import { ContainerPostDetails } from "./styled";
 
 const PostDetailsPage = () => {
     useProtectedPage();
-
+    
+    const params = useParams();
+    
     return (
-        <div>
+        <ContainerPostDetails>
             <h1>PostDetailsPage</h1>
-        </div>
+            <Post />
+        </ContainerPostDetails>
     )
 }
