+++ semana12/labeddit/src/pages/PostDetailsPage/styled.js
@@ -1,0 +1,10 @@
+import styled from "styled-components"
+
+export const ContainerPostDetails = styled.div`
+    font-family: 'Poppins', sans-serif;
+    display: grid;
+    justify-content: center;
+    align-items: center;
+    row-gap: 10px;
+    margin-bottom: 12px;
+`
